+++ internal/helm2/service_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package helm
+package helm2
 
 import (
 	"context"
